provide a clearer error message in case Popen'ed executable is not executable
subprocess.Popen() will give a proper error message if the first
argument is an absolute path, but if it's a bare command name, it will
complain with a "Permission Denied" error.

This change provides a clear error every time.

If speed is proven to have an impact on this code, we could add
the "Logs.verbose > 0" precondition or add memoization.