removed disable-doxygen callback warning
inspection of the OptionParserContext versus OptionContext.
Both do rather the same! The necessity of having them both
is questionable as redundant code introduces bugs and makes
the unreadable waf-source even worse... :(.

Lots of time spent to see that the above warning could just be commented
out. But for now the 'solution' has moved into documentation.py.