Generate doxygen into a sub-folder of build
When no OUTPUT_DIRECTORY was set for doxygen, then the parent
directory of the doxyfile was used. If the doxyfile is in the
root-directory, then this was the build-directory itself, which
led to the complete build-directory (including all other build
artifacts) to be installed.

The OUTPUT_DIRECTORY set (if not given) now includes the name
of the doxyfile itself (+ suffix '.doxy').

The install of doxygen-generated files also did not preserve
the directory structure.

The doxy playground example was simplified and updated, as
separate installation is no longer needed.

Signed-off-by: Thomas Nagy <tnagy2pow10@gmail.com>