Fix wrong use of Task.SKIP_ME in test_base
Skiping a task that was never run before by returning SKIP_ME in
runnable_status caused
"task in 'gtest-main' was not executed for some reason" exceptions.
It appears that SKIP_ME is only intended for task that doesn't need to
rerun.

This was triggered mainly by tests using the skip_run option

Change-Id: I44ced1b3fa820e8227b418a8f1dd4bfa43021e88