Fix installation of '.pyc' and '.pyo' files
This avoids issue #2067. This is a rare waf install bug, that occures
when '.pyc' files are found in the source tree. Changing the cwd of
the install task to the folder containing the from waf located '.pyc'
file fixes this (pyobj.parent instead of node.parent.get_bld()). 

As noted in the issues the exact cause of the bug is still unkown and
we can provoke it yet.



Change-Id: I5dd4d6231fe6c303643c0dd8cd42426bac2eabe1