Fix RPATH handling for some mpicc
This is the origin of the bug, waf assumes here that the `pc` file (or 
the mpi compiler output in this case) uses a certain format for `rpath` 
flags: `-Wl,-rpath /somepath` or `-Wl,-rpath,/somepath`. The code for 
the latter works and is just some lines below my change. However the 
code for the first case does not work if the package config contains 
something like `-Wl,-parth -Wl,/somepath` as it is transformed to 
`-Wl,-path,-Wl,/somepath` => the linker sees a `-Wl` which is wrong. 
My fix here just looks for the "wrong" assembly by checking for `-Wl,` 
in the next argument and stripping this string.
We could improvide this code by checking if the substring following the
`-Wl,` looks like a valid path (but this is difficult as paths may start
with `-` and all kind of things that also look like options).

A short walk through:
```
$ mpicc --showme:link
-pthread -L/somepath -Wl,-rpath -Wl,/somepath -lmpi
```

And here it is used:
```
$ cat ester/mpi/wscript 
# ...
    # check MPI (works only for OpenMPI)
    cfg.check_cfg(path = 'mpicxx', args = '--showme:compile',
            package = '', uselib_store = 'MPI4MPICONFIG')
    cfg.check_cfg(path = 'mpicxx', args = '--showme:link',
            package = '', uselib_store = 'MPI4MPICONFIG')
```

And `check_cfg` calls `exec_cfg` calls `parse_flags`.
Without this change `RPATH_MPI4MPICONFIG` contains `-Wl,/somepath` and 
is used via `RPATH_ST` (which is `-Wl,-rpath,%s`) => meeep, wrong.
With this change `RPATH_MPI4MPICONFIG` contains `/somepath` and the 
linker is called using `-Wl,-rpath,/somepath`, which is correct :).

Change-Id: I85bc831861a46be8a0f4d8d6a22c6a795336a427