Allow using directories as waf tools
Sometimes it is useful to be able to add a module to waf as a tool.

Using this patch one can use ./waf-light configure build --tools /tmp/mytool

This will add the files under /tmp/mytool under /waflib/extras/mytool. Such
that they can be imported in a wscript as from waflib.extras import mytool.