Recursively kill test child processes upon timeout
Previously we only killed the "direct" child of waf's test machinery.
This changeset introduces a heuristic to find the tree of all children
and tries to kill them all (starting with the tree leaves).
It first sends a SIGTERM to all children, then a SIGKILL.
The last process handled is the "direct" child.

Change-Id: Icc18ab64928c6b8b917ddc499acb826ff5e967fe