Improve gerrit changeset resolution
Currently, if there is more than one changeset for a project, the
first changeset is checked out and all others are cherry-picked. This
leads to errors in the following case:

ProjectX: review/master -> A -> B
                            `----`--- Depends-On: C
ProjectY: review/master -> C
                            `--- Depends-On: A

In the current setup, when we want to check out B, waf detects B's
dependency on C, checks out C. Then it discovers C's dependency on
A and tries to checkout A in ProjectX which - with the current
strategy - amounts to cherry-picking A on B which will fail in
almost any case.

The solution is to discover that the already checked out B is actually
a descendant of A - so by keeping B checked out we also meet the
requirement of A being checked out.

The full strategy is:
* check if one changeset is an ancestor of current HEAD or vice versa
  -> switch to the "younger" commit
* otherwise:
  * checkout on first changeset
  * cherry-pick the current changeset onto HEAD otherwise

Unfortunately, this strategy needs to be implemented as one single
bash command that is executed by mr.

Change-Id: Id2952fecea0757951502c6e7f943c04125a71b52