Request specifc patchlevel in Depends-On
This commit adds the ability to specify explicit patchlevels for
dependencies.
This is helpful when changes have major changes that are
not yet stable/time-efficient to rebase on.

Explicit patchlevels are specified with a slash after the change-number.
Requesting patchlevels 4{2,3,4} from changes 1234{5,6,7} but depending
ont he latest of changeset 12121 therefore becomes:

> Depends-On: 12345/42,12346/43,12347/44,12121

Please note that depending on specific patchlevels will most certainly
result in cherry-picks being applied when the commit-stack is
constructed in CI. The order in which cherry-picks are applied is the
order in which they are specified.

Change-Id: Idf0327e8ee60780b7a5f251893b52078bff03701