Add automatic dependency resolution for parents
By default waf will no walk up the commit tree and collect `Depends-On:`
specifications for all parents that have open changesets in gerrit.

In the case that this is not desired and all dependencies need to be
expressed explicitly (maybe even at a certain patchset level), toplevel
commit can specify:

> No-Parent-Depends-On

on a single line in their commit message in order to disable patch-set
resolutin for parents. Their dependencies will also _not_ have their
parents scanned for Depends-On-specifications. Via this one can use one
commit ontop of one project to build his/her entire commit-stack in all
projects out of hand-picked changes/patchlevels that are cherry-picked
ontop each other. This is expected to be the exception rather than the
norm!

Currently, encountering a `No-Parent-Depends-On` in the middle of
another project-stack while resolving dependencies will be silently
ignored (but will probably lead to a build error). It should always be
preferable to rebase ontop of current-patchsets in all stacks so that
`No-Parent-Depends-On` will not be needed.

Change-Id: Iefce885615d5bbd2c3f2ae7a1db853afe0a0a473