Improve mixing of multiple commit stacks
When one wishes to check out multiple commit stack (e.g. in the current
case a hashtag and a specific other change on top of a stack of changes)
it can happen that some changes are attempted to be cherry picked
multiple times because when waf cannot establish a clear ancestral
relationship it resorts to cherry-picking. In most cases this fails.

One remedy is to check if the current change-to-be-applied was already
cherry-picked by looking for its "Change-Id" in the git log of the
current stack.

Even for huge git logs the slowdown on check-out should be manageable.

This will of course fail if both commit stacks are not trivially
cherry-pickable on top of each other, same as before.

Depends-On: 12055,12062

Change-Id: Ied595204681e9c28a035c5f556b2368b84c0bdbd