Check if change-id is present prior to applying
With the newest changes in regards to mixing multiple stacks, there are
a lot of changes to apply to any one project. We can speed this up by
first checking if the given change-id is already present in the git log
and only _then_ performing the fetch if it is not.

Change-Id: I6604083120b1942b4902631b4c3af68d30b5609d